fix warnings and bug in shaderloading

+++ Code/Renderer/Shaders/ShaderLoading.cpp
@@ -118,7 +118,7 @@ std::span<std::span<wchar_t>> parseShaderIncludeStrings(MemoryArena::memoryArena
 void copySubstring(std::span<wchar_t> sourceA, std::span<wchar_t> sourceB, std::span<wchar_t> tgt)
 {
     assert (tgt.size() <= sourceA.size() + sourceB.size());
-    uint32_t headLength = sourceA.size();
+    size_t headLength = sourceA.size();
     //Copy head
     for (int j = 0; j < headLength; j++)
     {
@@ -142,7 +142,7 @@ std::span<std::span<wchar_t>>  findShaderIncludes(MemoryArena::memoryArena* allo
 {
    
     size_t filenameStart = 0;
-    for(int i = shaderPath.length();i > 0; i --)
+    for(int i = (int)shaderPath.length();i > 0; i --)
     {
         if (shaderPath[i] == L'\\' || shaderPath[i] == L'/') //At last delimeter
         {
@@ -253,7 +253,7 @@ struct loadedBlob
 
     loadedBlob(SIZE_T size, std::unique_ptr<uint32_t[]> _buffer)
     {
-        this->size = size; //TODO JS: magic number i dont understand
+        this->size = (uint32_t)size; //TODO JS: magic number i dont understand
         this->buffer = std::move(_buffer);
     }
 };
